+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - added command banner:  display an awesome wp3 banner 
 - added many improvements into system modules
 - added improves module for running in background
+- added command kill: terminate a module in background by id 
 
 ### Changed
 - moved command info to extensions directory